refactor(app): BaseMviActivity.kt, BaseMviFragment.kt added.
