--- app/src/main/java/com/github/mvibase/base/BaseMviViewEvent.kt
@@ -1,3 +1,0 @@
-package com.github.mvibase.base
-
-interface BaseMviViewEvent
