+++ app/src/main/java/com/github/mvibase/presentation/BaseMviAction.kt
@@ -1,4 +1,4 @@
-package com.github.mvibase.base
+package com.github.mvibase.presentation
 
 
 /**
