+++ app/src/main/java/com/github/mvibase/presentation/BaseMviIntent.kt
@@ -1,4 +1,4 @@
-package com.github.mvibase.base
+package com.github.mvibase.presentation
 
 /**
  * Immutable object which represent an view's intent.
