+++ app/src/main/java/com/github/mvibase/presentation/BaseMviViewEvent.kt
@@ -1,0 +1,3 @@
+package com.github.mvibase.presentation
+
+interface BaseMviViewEvent
