+++ app/src/main/java/com/github/mvibase/presentation/BaseMviViewState.kt
@@ -1,4 +1,4 @@
-package com.github.mvibase.base
+package com.github.mvibase.presentation
 
 
 /**
