+++ app/src/main/java/com/github/mvibase/view/BaseMviFragment.kt
@@ -1,0 +1,32 @@
+package com.github.mvibase.view
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+
+abstract class BaseMviFragment<VB : ViewDataBinding> : Fragment() {
+
+    lateinit var dataBinding: VB
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        dataBinding = DataBindingUtil.inflate(
+            LayoutInflater.from(context),
+            getResourceLayoutId(),
+            container,
+            false
+        )
+
+        return dataBinding.root
+    }
+
+    abstract fun getResourceLayoutId(): Int
+
+}
